backend/requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
websockets